Validate image location scheme
This fixes the other half of
https://bugs.launchpad.net/glance/+bug/1400966 the directory traversal
bug. In havana, glance would prevent you part of the flaw, by way of
https://bugs.launchpad.net/glance/+bug/942118 but glance would still
allow user to update an image location after an image was created
without validation of the image scheme. This patch closes the loop hole.

This also adds a test to ensure these protected schemes are not allowed
to be used when updating an image location.